+++ analyze.py
@@ -4,6 +4,8 @@ from sympy import isprime
 from enum import Enum
 import os
 
+from configuration import NUMBERS
+
 def pretty_print(f):
 	def wrapped(*args, **kw):
 		n = len(f.__name__)
@@ -24,10 +26,12 @@ class Analyzer:
 		self.format_data()
 
 	def load_data(self):
-		dat_filename = "output.dat"
-		csv_filename = "output.csv"
+		dat_filename = f"output_{NUM_NUMBERS}.dat"
+		csv_filename = f"output_{NUM_NUMBERS}.csv"
 		if os.path.isfile(dat_filename):
 			data = np.fromfile(dat_filename, dtype=np.int32)
+			n = data.shape[0]
+			data.shape = (n // (NUM_NUMBERS + 1000), NUM_NUMBERS + 1000)
 		elif os.path.isfile(csv_filename):
 			data = np.loadtxt(open(csv_filename, "rb"), 
 				delimiter=",", skiprows=0, dtype=np.int32)
@@ -39,12 +43,10 @@ class Analyzer:
 
 	def format_data(self):
 		data = self.load_data()
-		n = data.shape[0]
-		data.shape = (n // 1006, 1006)
 		self.offset = 0 if self.include else 100
 
-		self.boards = data[:,:6]
-		self.counts = data[:,6+self.offset:]
+		self.boards = data[:,:NUM_NUMBERS]
+		self.counts = data[:,NUM_NUMBERS+self.offset:]
 		self.numbers = np.arange(self.offset, 1000)
 		self.numbers_per_board = (self.counts[:,:] > 0).sum(axis=1)
 		self.boards_per_number = (self.counts[:,:] > 0).sum(axis=0)
@@ -93,7 +95,7 @@ class Analyzer:
 		print("Distribution of high digits (25-100) among these boards:")
 		for i, n in enumerate(high_digits_hist):
 			print(f"    {i} high digits: {len(n)} boards")
-			if i in (0, 3):
+			if len(high_digits_hist[i]) < 10:
 				for eb in high_digits_hist[i]:
 					print(" "*8 + str(eb))
