+++ configuration.py
@@ -1,0 +1,10 @@
+NUM_NUMBERS = 6
+NUM_SYMBOLS = NUM_NUMBERS - 1
+NUM_TOKENS = NUM_NUMBERS + NUM_SYMBOLS
+MAX_TARGET = 1000
+
+NUM_EXTRA_VALUES = 4
+SUBTRACTION_FAIL_INDEX = 0
+DIVISION_FAIL_INDEX = 1
+PERMUTATION_FAIL_INDEX = 2
+PERMUTATION_SUCCESS_INDEX = 3
