+++ countdown_kernel.cl
@@ -1,12 +1,13 @@
-#define NUM_TOKENS 11
-#define MAX_TARGET 1000
 #define RPN_STACK_SIZE 10
 
-#define NUM_EXTRA_VALUES 4
-#define SUBTRACTION_FAIL_INDEX 1000
-#define DIVISION_FAIL_INDEX 1001
-#define PERMUTATION_FAIL_INDEX 1002
-#define PERMUTATION_SUCCESS_INDEX 1003
+typedef int Tokens[NUM_TOKENS];
+typedef int Counts[MAX_TARGET];
+typedef int ExtraStats[NUM_EXTRA_VALUES];
+
+struct Result {
+	Counts counts;
+	ExtraStats extra_stats;
+};
 
 void swap(int* a, int* b) {
 	int tmp = *a;
@@ -15,8 +16,9 @@ void swap(int* a, int* b) {
 }
 
 void reverse(int* first, int* last) {	
-	for (; first != last && first != --last; ++first)
+	for (; first != last && first != --last; ++first) {
 		swap(first, last);
+	}
 }
 
 void next_permutation(int* first, int* last) {
@@ -26,23 +28,19 @@ void next_permutation(int* first, int* last) {
 		return;
 	}
 
-	while(true)
-	{
+	while(true) {
 		int* next1 = next;
 		--next;
-		if(*next < *next1)
-		{
+		if(*next < *next1) {
 			int* mid = last;
 			--mid;
-			for(; !(*next < *mid); --mid)
-				;
+			for(; !(*next < *mid); --mid);
 			swap(next, mid);
 			reverse(next1, last);
 			return;
 		}
 
-		if(next == first)
-		{
+		if(next == first) {
 			reverse(first, last);
 			return;
 		}
@@ -56,8 +54,8 @@ void print(int nums[], int n) {
 	printf("\n");
 }
 
-void evaluate_rpn(int tokens[NUM_TOKENS], int res[RPN_STACK_SIZE], int *res_n, 
-	int local_result[MAX_TARGET + NUM_EXTRA_VALUES]) {
+void evaluate_rpn(Tokens tokens, int res[RPN_STACK_SIZE], int *res_n, 
+	struct Result *local_result) {
 
 	int stack[RPN_STACK_SIZE];
 	int stack_i = 0;
@@ -79,7 +77,7 @@ void evaluate_rpn(int tokens[NUM_TOKENS], int res[RPN_STACK_SIZE], int *res_n,
 				case -2: // subtraction
 					if (b > a) {
 						*res_n = res_i;
-						local_result[SUBTRACTION_FAIL_INDEX]++;
+						local_result->extra_stats[SUBTRACTION_FAIL_INDEX]++;
 						return;
 					}
 					val = a - b;
@@ -90,7 +88,7 @@ void evaluate_rpn(int tokens[NUM_TOKENS], int res[RPN_STACK_SIZE], int *res_n,
 				case -4: // division
 					if (b == 0 || a % b != 0) {
 						*res_n = res_i;
-						local_result[DIVISION_FAIL_INDEX]++;
+						local_result->extra_stats[DIVISION_FAIL_INDEX]++;
 						return;
 					}
 					val = a / b;
@@ -119,17 +117,15 @@ bool check_rpn(int tokens[NUM_TOKENS]) {
 	return true;
 }
 
-kernel void evaluate(__global int *data_g, __global int *result, 
+kernel void evaluate(
+	__global Tokens *data_g, __global struct Result *result, 
 	__constant int *dims) {
 	
 	int tid = get_global_id(0);
-	int idx = dims[1] * tid;
-	int local_data[NUM_TOKENS];
+	Tokens local_data;
 	for (int i = 0; i < NUM_TOKENS; i++) {
-		local_data[i] = data_g[i + idx];
+		local_data[i] = data_g[tid][i];
 	}
-	// print(local_data, NUM_TOKENS);
-	// printf("%d %d %d\n", dims[0], dims[1], dims[2]);
 	
 	int *first = &local_data[0];
 	int *last = &local_data[NUM_TOKENS];
@@ -137,29 +133,35 @@ kernel void evaluate(__global int *data_g, __global int *result,
 	int rpn_res[RPN_STACK_SIZE];
 	int rpn_n;
 	
-	int local_result[MAX_TARGET + NUM_EXTRA_VALUES];
-	for (int i = 0; i < MAX_TARGET + NUM_EXTRA_VALUES; i++) {
-		local_result[i] = 0;
+	struct Result local_result;
+	for (int i = 0; i < MAX_TARGET; i++) {
+		local_result.counts[i] = 0;
+	}
+	for (int i = 0; i < NUM_EXTRA_VALUES; i++) {
+		local_result.extra_stats[i] = 0;
 	}
 
 	for (int i = 0; i < limit; i++) {
 		bool check = check_rpn(local_data);
 		if (check) {
-			evaluate_rpn(local_data, rpn_res, &rpn_n, local_result);
+			evaluate_rpn(local_data, rpn_res, &rpn_n, &local_result);
 			for (int j = 0; j < rpn_n; j++) {
 				int val = rpn_res[j];
 				int in_range = 0 <= val & val < MAX_TARGET;
-				local_result[val * in_range] += in_range;
+				local_result.counts[val * in_range] += in_range;
 			}
-			local_result[PERMUTATION_SUCCESS_INDEX]++;
+			local_result.extra_stats[PERMUTATION_SUCCESS_INDEX]++;
 		} else {
-			local_result[PERMUTATION_FAIL_INDEX]++;
+			local_result.extra_stats[PERMUTATION_FAIL_INDEX]++;
 		}
 		next_permutation(first, last);
 	}
 
-	for (int i = 0; i < MAX_TARGET + NUM_EXTRA_VALUES; i++) {
-		result[(MAX_TARGET + NUM_EXTRA_VALUES) * tid + i] = local_result[i];
+	for (int i = 0; i < MAX_TARGET; i++) {
+		result[tid].counts[i] = local_result.counts[i];
+	}
+	for (int i = 0; i < NUM_EXTRA_VALUES; i++) {
+		result[tid].extra_stats[i] = local_result.extra_stats[i];
 	}
 
 }
